// frontend/src/app/models/korisnik.ts
export class Korisnik{
    korisnickoIme: string = '';
    lozinka: string = '';
    ime: string = '';
    prezime: string = '';
    adresa: string = '';
    telefon: string = '';
    mejl: string = '';
    tip: string = '';
    brojLekLic: string = '';
    specijalizacija: string = '';
    ogranak: string = '';
    vrstePregleda: Array<string>;
    status: string = '';
    slika: string = '';
    slobodniDani: Array<string>;
    godisnji: Array<string>;
}